Dagger modules refactored
