+++ hedera-node/hedera-app/src/main/java/com/hedera/node/app/HederaApp.java
@@ -19,35 +19,18 @@ package com.hedera.node.app;
 import com.hedera.hapi.node.base.AccountID;
 import com.hedera.node.app.annotations.MaxSignedTxnSize;
 import com.hedera.node.app.annotations.NodeSelfId;
-import com.hedera.node.app.authorization.AuthorizerDaggerModule;
+import com.hedera.node.app.authorization.AuthorizerModule;
 import com.hedera.node.app.components.IngestComponent;
 import com.hedera.node.app.components.QueryComponent;
 import com.hedera.node.app.fees.FeesModule;
-import com.hedera.node.app.info.InfoDaggerModule;
-import com.hedera.node.app.metrics.MetricsDaggerModule;
+import com.hedera.node.app.info.InfoModule;
+import com.hedera.node.app.metrics.MetricsModule;
+import com.hedera.node.app.service.mono.LegacyMonoModule;
 import com.hedera.node.app.service.mono.ServicesApp;
-import com.hedera.node.app.service.mono.config.ConfigModule;
-import com.hedera.node.app.service.mono.context.ContextModule;
 import com.hedera.node.app.service.mono.context.annotations.BootstrapProps;
 import com.hedera.node.app.service.mono.context.annotations.StaticAccountMemo;
-import com.hedera.node.app.service.mono.context.properties.PropertiesModule;
 import com.hedera.node.app.service.mono.context.properties.PropertySource;
-import com.hedera.node.app.service.mono.contracts.ContractsModule;
-import com.hedera.node.app.service.mono.files.FilesModule;
-import com.hedera.node.app.service.mono.grpc.GrpcModule;
-import com.hedera.node.app.service.mono.keys.KeysModule;
-import com.hedera.node.app.service.mono.ledger.LedgerModule;
-import com.hedera.node.app.service.mono.queries.QueriesModule;
-import com.hedera.node.app.service.mono.records.RecordsModule;
-import com.hedera.node.app.service.mono.sigs.SigsModule;
 import com.hedera.node.app.service.mono.state.StateModule;
-import com.hedera.node.app.service.mono.state.expiry.ExpiryModule;
-import com.hedera.node.app.service.mono.state.tasks.TaskModule;
-import com.hedera.node.app.service.mono.stats.StatsModule;
-import com.hedera.node.app.service.mono.store.StoresModule;
-import com.hedera.node.app.service.mono.throttling.ThrottlingModule;
-import com.hedera.node.app.service.mono.txns.TransactionsModule;
-import com.hedera.node.app.service.mono.txns.submission.SubmissionModule;
 import com.hedera.node.app.service.mono.utils.NonAtomicReference;
 import com.hedera.node.app.services.ServicesModule;
 import com.hedera.node.app.solvency.SolvencyModule;
@@ -55,11 +38,8 @@ import com.hedera.node.app.state.HederaState;
 import com.hedera.node.app.state.HederaStateModule;
 import com.hedera.node.app.state.WorkingStateAccessor;
 import com.hedera.node.app.throttle.ThrottleModule;
-import com.hedera.node.app.workflows.handle.HandleWorkflowModule;
-import com.hedera.node.app.workflows.ingest.IngestModule;
+import com.hedera.node.app.workflows.WorkflowsModule;
 import com.hedera.node.app.workflows.prehandle.AdaptedMonoEventExpansion;
-import com.hedera.node.app.workflows.prehandle.PreHandleWorkflowModule;
-import com.hedera.node.app.workflows.query.QueryWorkflowModule;
 import com.swirlds.common.crypto.Cryptography;
 import com.swirlds.common.crypto.Hash;
 import com.swirlds.common.system.InitTrigger;
@@ -78,41 +58,16 @@ import javax.inject.Singleton;
 @Singleton
 @Component(
         modules = {
-            TaskModule.class,
-            com.hedera.node.app.service.mono.fees.FeesModule.class,
-            KeysModule.class,
-            SigsModule.class,
-            GrpcModule.class,
-            ConfigModule.class,
-            StatsModule.class,
-            StateModule.class,
-            FilesModule.class,
-            LedgerModule.class,
-            StoresModule.class,
-            ContextModule.class,
-            RecordsModule.class,
-            QueriesModule.class,
-            ContractsModule.class,
-            PropertiesModule.class,
-            ThrottlingModule.class,
-            SubmissionModule.class,
-            TransactionsModule.class,
-            ExpiryModule.class,
-            ServicesModule.class,
-            IngestModule.class,
-            QueryWorkflowModule.class,
-            HandleWorkflowModule.class,
-            PreHandleWorkflowModule.class,
-            HederaStateModule.class,
-            FeesModule.class,
-            HederaStateModule.class,
-            MetricsDaggerModule.class,
-            AuthorizerDaggerModule.class,
-            InfoDaggerModule.class,
-            ThrottleModule.class,
-            SolvencyModule.class,
-            HandleWorkflowModule.class,
-            ConfigModule.class
+                LegacyMonoModule.class,
+                ServicesModule.class,
+                WorkflowsModule.class,
+                HederaStateModule.class,
+                FeesModule.class,
+                MetricsModule.class,
+                AuthorizerModule.class,
+                InfoModule.class,
+                ThrottleModule.class,
+                SolvencyModule.class
         })
 public interface HederaApp extends ServicesApp {
     /* Needed by ServicesState */
