+++ hedera-node/hedera-app/src/main/java/com/hedera/node/app/authorization/AuthorizerModule.java
@@ -14,17 +14,20 @@
  * limitations under the License.
  */
 
-package com.hedera.node.app.workflows.ingest;
+package com.hedera.node.app.authorization;
 
-import com.hedera.node.app.components.IngestComponent;
-import dagger.Binds;
+import com.hedera.node.app.service.mono.context.domain.security.HapiOpPermissions;
 import dagger.Module;
+import dagger.Provides;
+import edu.umd.cs.findbugs.annotations.NonNull;
+import javax.inject.Singleton;
 
-/**
- * Module for Ingest processing.
- */
-@Module(subcomponents = {IngestComponent.class})
-public interface IngestModule {
-    @Binds
-    IngestWorkflow bindIngestWorkflow(IngestWorkflowImpl ingestWorkflow);
+/** A Dagger module for providing dependencies based on {@link Authorizer}. */
+@Module
+public interface AuthorizerModule {
+    @Provides
+    @Singleton
+    static Authorizer provideAuthorizer(@NonNull final HapiOpPermissions permissions) {
+        return new AuthorizerImpl(permissions);
+    }
 }
