+++ hedera-node/hedera-app/src/main/java/com/hedera/node/app/workflows/WorkflowsModule.java
@@ -1,0 +1,36 @@
+/*
+ * Copyright (C) 2023 Hedera Hashgraph, LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+package com.hedera.node.app.workflows;
+
+import com.hedera.node.app.workflows.handle.HandleWorkflowModule;
+import com.hedera.node.app.workflows.ingest.IngestWorkflowModule;
+import com.hedera.node.app.workflows.prehandle.PreHandleWorkflowModule;
+import com.hedera.node.app.workflows.query.QueryWorkflowModule;
+import dagger.Module;
+
+/**
+ * Dagger module for all workflows
+ */
+@Module(includes = {
+        HandleWorkflowModule.class,
+        IngestWorkflowModule.class,
+        PreHandleWorkflowModule.class,
+        QueryWorkflowModule.class
+})
+public interface WorkflowsModule {
+}
